Move script test for d-c-p to test_scripts.py
