+++ tests/test_scripts.py
@@ -1,4 +1,7 @@
+import re
+from itertools import product
 from datetime import datetime
+from typing import List
 
 import pytest
 import pytz
@@ -10,6 +13,7 @@ from crmprtd.download_utils import verify_date
 # Must import the module, not objects from it, so we can mock the objects.
 import crmprtd.process
 import crmprtd.download
+import crmprtd.download_cache_process
 
 
 @pytest.mark.parametrize(
@@ -107,3 +111,182 @@ def test_process_main(
         is_diagnostic=diagnostic,
         do_infer=infer,
     )
+
+
+def match_args(arglist: List[str], pattern: str):
+    return re.match(pattern, " ".join(arglist))
+
+
+def assert_match_args(*args):
+    assert match_args(*args) is not None
+
+
+@pytest.mark.parametrize(
+    "network, tag, frequency, log_filename, cache_filename, connection_string, "
+    "seq_of_expected_cmds_patterns, expected_final_destination_pattern",
+    # This covers all current networks and aliases.
+    [
+        # Tests of all networks, with default log/cache filenames
+        (
+            "bc_hydro",
+            "tag",
+            None,
+            None,
+            None,
+            None,
+            [
+                [
+                    r"crmprtd_download -N bc_hydro",
+                ],
+            ],
+            r"~/bc_hydro/cache/tag_bc_hydro_.*\.txt",
+        ),
+        *(
+            (
+                network,
+                "tag",
+                None,
+                None,
+                None,
+                "DSN",
+                [
+                    [
+                        rf"crmprtd_download -N {network}",
+                        rf"tee\s+~/{network}/cache/tag_{network}_.*.txt",
+                        rf"crmprtd_process -N {network} -c DSN",
+                    ],
+                ],
+                None,
+            )
+            for network in "crd moti wamr wmb".split()
+        ),
+        (
+            "ec",
+            "tag",
+            "daily",
+            None,
+            None,
+            "DSN",
+            [
+                [
+                    rf"crmprtd_download -N ec -p {prov} -F daily",
+                    rf"tee\s+~/ec/cache/tag_daily_{prov}_.*.xml",
+                    r"crmprtd_process -N ec -c DSN",
+                ]
+                for prov in "bc yt".split()
+            ],
+            None,
+        ),
+        *(
+            (
+                alias,
+                "tag",
+                None,
+                None,
+                None,
+                "DSN",
+                [
+                    [
+                        rf"crmprtd_download -N {network}",
+                        rf"tee\s+~/{network}/cache/tag_{network}_.*.xml",
+                        rf"crmprtd_process -N {network} -c DSN",
+                    ]
+                    for network in crmprtd.download_cache_process.network_aliases[alias]
+                ],
+                None,
+            )
+            for alias in "hourly_swobml2 ytnt".split()
+        ),
+        # Tests of output filename handling
+        *(
+            (
+                network,
+                "tag",
+                None,
+                log_filename,
+                cache_filename,
+                "DSN",
+                [
+                    [
+                        (
+                            rf"crmprtd_download -N {network}"
+                            rf" .*--log_filename {log_filename or f'~/{network}/logs/tag_{network}_json.log'}"
+                        ),
+                        rf"tee\s+{cache_filename or rf'~/{network}/cache/tag_{network}_.*.txt'}",
+                        rf"crmprtd_process -N {network} -c DSN",
+                    ],
+                ],
+                None,
+            )
+            for network, log_filename, cache_filename in product(
+                "crd".split(),  # Probably only need to check on one network
+                (None, "custom_log_filename"),
+                (None, "custom_cache_filename"),
+            )
+        ),
+    ],
+)
+def test_download_cache_process_main(
+    network,
+    tag,
+    frequency,
+    log_filename,
+    cache_filename,
+    connection_string,
+    seq_of_expected_cmds_patterns,
+    expected_final_destination_pattern,
+    mocker,
+):
+    arglist = f"-N {network}"
+    if tag is not None:
+        arglist += f" -T {tag}"
+    if frequency is not None:
+        arglist += f" -F {frequency}"
+    if log_filename is not None:
+        arglist += f" --log_filename {log_filename}"
+    if cache_filename is not None:
+        arglist += f" --cache_filename {cache_filename}"
+    if connection_string is not None:
+        arglist += f" -c {connection_string}"
+
+    # In certain cases, the unit under test will try to open a file.
+    # Instead, don't open a file but return the filepath for checking.
+    def new_open(filepath, *args, **kwargs):
+        return filepath
+
+    mocker.patch("builtins.open", new=new_open)
+
+    # Mock chain_subprocesses, so that we can see what it is being called with. That's
+    # the evidence we use to certify that the script is working as expected.
+    # chain_subprocesses itself is tested separately.
+    cp = mocker.patch("crmprtd.infill.chain_subprocesses", return_value=True)
+    crmprtd.download_cache_process.main(arglist.split())
+    call_args_list = cp.call_args_list
+
+    # Note: call_args_list is a list of the arguments of calls to chain_subprocesses
+    # from each element of that arguments list we extract the first argument
+    #   (which is a list of the commands to be chained)
+    # seq_of_expected_cmds_prefixes is a list of expected prefixes (initial elements)
+    #   of commands
+    # we compare actual commands to the prefixes.
+
+    # Check that the expected number of calls was made.
+    assert len(call_args_list) == len(seq_of_expected_cmds_patterns)
+
+    # Check that each call contained the expected commands (or prefixes thereof).
+    for call_arg, expected_cmds_patterns in zip(
+        call_args_list, seq_of_expected_cmds_patterns
+    ):
+        commands, *_ = call_arg.args
+        final_destination = call_arg.kwargs["final_destination"]
+        # Check that the number of commands in the call is as expected.
+        assert len(commands) == len(expected_cmds_patterns)
+        # Check that each command matches the expected pattern.
+        for cmd, expected_cmd_pattern in zip(commands, expected_cmds_patterns):
+            assert_match_args(cmd, expected_cmd_pattern)
+        # Check that final destination is as expected.
+        assert final_destination is None or re.match(
+            expected_final_destination_pattern, final_destination
+        )
+
+    cp.reset_mock()
